perf(frontend): implement route-level code splitting and lazy load stellar-sdk
