+++ frontend/src/utils/validation.ts
@@ -1,12 +1,78 @@
 // Validation utilities
-import { StrKey } from 'stellar-sdk'
+const ALPHABET = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ234567'
+
+function base32Decode(input: string): Uint8Array {
+  const upper = input.toUpperCase().replace(/=+$/, '')
+  const length = upper.length
+  const count = Math.floor((length * 5) / 8)
+  const result = new Uint8Array(count)
+
+  let buffer = 0
+  let bitsLeft = 0
+  let next = 0
+
+  for (let i = 0; i < length; i++) {
+    const val = ALPHABET.indexOf(upper[i])
+    if (val === -1) throw new Error('Invalid base32 character')
+    buffer = (buffer << 5) | val
+    bitsLeft += 5
+    if (bitsLeft >= 8) {
+      result[next++] = (buffer >> (bitsLeft - 8)) & 0xff
+      bitsLeft -= 8
+    }
+  }
+  return result
+}
+
+function crc16(data: Uint8Array): number {
+  let crc = 0x0000
+  for (let i = 0; i < data.length; i++) {
+    const byte = data[i]
+    let code = (crc >>> 8) & 0xff
+    code ^= byte
+    code ^= code >>> 4
+    crc = (crc << 8) ^ (code << 12) ^ (code << 5) ^ code
+    crc &= 0xffff
+  }
+  return crc
+}
 
 export const isValidStellarAddress = (address: string): boolean => {
-  return StrKey.isValidEd25519PublicKey(address)
+  try {
+    if (address.length !== 56) return false
+    if (address[0] !== 'G') return false
+    const decoded = base32Decode(address)
+    if (decoded.length !== 35) return false
+    const versionByte = decoded[0]
+    if (versionByte !== 0x30) return false // 6 << 3 = 48 (0x30) for Ed25519 Public Key
+
+    const payload = decoded.slice(1, 33)
+    const checksum = decoded.slice(33, 35)
+    const calculatedCrc = crc16(new Uint8Array([versionByte, ...payload]))
+    const expectedCrc = checksum[0] | (checksum[1] << 8)
+    return calculatedCrc === expectedCrc
+  } catch {
+    return false
+  }
 }
 
 export const isValidContractAddress = (address: string): boolean => {
-  return StrKey.isValidContract(address)
+  try {
+    if (address.length !== 56) return false
+    if (address[0] !== 'C') return false
+    const decoded = base32Decode(address)
+    if (decoded.length !== 35) return false
+    const versionByte = decoded[0]
+    if (versionByte !== 0x10) return false // 2 << 3 = 16 (0x10) for Contract
+
+    const payload = decoded.slice(1, 33)
+    const checksum = decoded.slice(33, 35)
+    const calculatedCrc = crc16(new Uint8Array([versionByte, ...payload]))
+    const expectedCrc = checksum[0] | (checksum[1] << 8)
+    return calculatedCrc === expectedCrc
+  } catch {
+    return false
+  }
 }
 
 export const validateTokenParams = (params: {
